+++ global/types/global.d.ts
@@ -4,6 +4,7 @@ declare global {
         interface ProcessEnv {
             readonly NODE_ENV: 'development' | 'production' | 'test';
             readonly PORT: string;
+            readonly MONGO_URI: string;
         }
     }
 }
