+++ server/src/api/v2/routes/tarot.routes.ts
@@ -3,8 +3,7 @@ import tarotController from '../controllers/tarot.controller';
 
 const router = Router();
 
-router.route('/').get(tarotController.getAll);
-router.route('/new').get(tarotController.getAllNew);
+router.route('/').get(tarotController.getAllNew);
 
 router.route('/img/:suit/:rank').get(tarotController.sendNewImage);
